+++ src/libswf/tag.c
@@ -38,9 +38,9 @@ void tagShowFrame(SWFAppContext* app_context)
 			case CHAR_TYPE_TEXT:
 				flashbang_upload_extra_transform_id(context, obj->transform_id);
 				flashbang_upload_cxform_id(context, ch->cxform_id);
-				for (size_t j = 0; j < ch->text_size; ++j)
+				for (u32 j = 0; j < ch->text_size; ++j)
 				{
-					size_t glyph_index = 2*app_context->text_data[ch->text_start + j];
+					u32 glyph_index = 2*app_context->text_data[ch->text_start + j];
 					flashbang_draw_shape(context, app_context->glyph_data[glyph_index], app_context->glyph_data[glyph_index + 1], ch->transform_start + j);
 				}
 				break;
@@ -50,7 +50,7 @@ void tagShowFrame(SWFAppContext* app_context)
 	flashbang_close_pass(context);
 }
 
-void tagDefineShape(SWFAppContext* app_context, CharacterType type, size_t char_id, size_t shape_offset, size_t shape_size)
+void tagDefineShape(SWFAppContext* app_context, CharacterType type, u32 char_id, u32 shape_offset, u32 shape_size)
 {
 	ENSURE_SIZE(dictionary, char_id, dictionary_capacity, sizeof(Character));
 
@@ -59,7 +59,7 @@ void tagDefineShape(SWFAppContext* app_context, CharacterType type, size_t char_
 	dictionary[char_id].size = shape_size;
 }
 
-void tagDefineText(SWFAppContext* app_context, size_t char_id, size_t text_start, size_t text_size, u32 transform_start, u32 cxform_id)
+void tagDefineText(SWFAppContext* app_context, u32 char_id, u32 text_start, u32 text_size, u32 transform_start, u32 cxform_id)
 {
 	ENSURE_SIZE(dictionary, char_id, dictionary_capacity, sizeof(Character));
 
@@ -70,7 +70,7 @@ void tagDefineText(SWFAppContext* app_context, size_t char_id, size_t text_start
 	dictionary[char_id].cxform_id = cxform_id;
 }
 
-void tagPlaceObject2(SWFAppContext* app_context, size_t depth, size_t char_id, u32 transform_id)
+void tagPlaceObject2(SWFAppContext* app_context, u32 depth, u32 char_id, u32 transform_id)
 {
 	ENSURE_SIZE(display_list, depth, display_list_capacity, sizeof(DisplayObject));
 
@@ -83,7 +83,7 @@ void tagPlaceObject2(SWFAppContext* app_context, size_t depth, size_t char_id, u
 	}
 }
 
-void defineBitmap(size_t offset, size_t size, u32 width, u32 height)
+void defineBitmap(u32 offset, u32 size, u32 width, u32 height)
 {
 	flashbang_upload_bitmap(context, offset, size, width, height);
 }
